# openapi.yaml
openapi: 3.0.3
info:
  title: Vaultora API
  version: 1.0.0
  description: Private self-hosted file storage. Session cookie `vid_session` + `X-CSRF-Token` header on mutating requests.
servers:
  - url: /
paths:
...
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                totp: { type: string }
      responses: { '200': { description: session }, '401': { description: invalid } }
...
  /api/users:
    get: { summary: List users (admin), responses: { '200': { description: users } } }
    post:
      summary: Create user (admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                role: { type: string, enum: [admin, user] }
                quota_bytes: { type: integer, nullable: true }
      responses: { '201': { description: created } }
...
  /api/files:
    get:
      summary: List folder
      parameters:
        - { name: path, in: query, schema: { type: string }, example: /My Files }
        - { name: sort, in: query, schema: { type: string, enum: [name, size, date, type] } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
      responses: { '200': { description: listing } }
...